requirements.txt
Django~=2.2.4
django-crispy-forms
python-slugify
